Method for commeting
Implmeneted a method to add a comment to a specific task ID. This is an
inital commit with only method implemented. Nothing else has been done
due to missing MainActivity Table.

To-Do:

Display Window as a popup
Link it to the button on the main table
Polish the interface
Add look at all comments feature